Basic miniuart support in l.S (raspberry pi)
Running r9 on qemu mimicking a raspberry pi 3 works as expected, writing out system info via the uart, but running on a real raspberry pi 4 does nothing. This PR enabled the miniuart in l.S, allowing us to write single characters to support basic debugging at the very early stages.

This MR also changes the way constants are used in aarch64. It will run on a raspberry pi 4 and write . to the miniuart. I haven't been able to get it running on rpi3. It still works as before on a qemu thinking it's an rpi3.

(Using this shows that r9 dies when enabling the mmu - that's the next problem)

Signed-off-by: Graham MacDonald <grahamamacdonald@gmail.com>